new spring system
Removed old system of constraining wheels with pin joints and two
springs, as well as a spring between the wheels.
This system uses two separate springs on each wheel going from the body
to the center of the wheel in a "V" shape as well as a spring between
the wheels to keep the wheels from getting pushed "up and under" the
truck.